+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 Notable changes to this project.
 
+## unreleased
+- add `round_details`, returning correlation scores of all users for the round
 
 ## [2.1.0] - 2019-11-15
 - add some more details to `get_leaderboard`